Modelling RT as secondary treatment for conservative management
Cure model giving the probability of not having radiation therapy
following assignment to conservative management (active surveillance &
watchfull waiting). Log-normal distribution of the time to radiation
therapy for those that do have that. A future extension would be to
model active surveillance and watchfull waiting separately.

Radiation therapy in the outpatient setting can be defined
(https://www.health.mil/Reference-Center/Publications/2015/09/01/Prostate-Cancer)
as Z510* for the primary diagnosis and C61 as the secondary diagnosis.
Taking this definition, the probability of an RT and then an RP is
roughly similar to an RP alone, while the probability of an RP followed
by RT is very low. As a simplification, we can simulate separately for
RP and RT and remove the RT if it follows the RP. This means that a man
can have an RT and then an RP (both not an RP then an RT).